chore: update dependencies to latest versions
- rusqlite: 0.32 -> 0.38 (fix usize -> i64 cast for ToSql)
- zip: 2.2 -> 7.2
- pulldown-cmark: 0.12 -> 0.13
- toml: 0.8 -> 0.9

All tests passing, no breaking API changes in our code.

Closes #49

+++ crates/ankit-builder/src/apkg.rs
@@ -151,7 +151,7 @@ impl ApkgBuilder {
                 conn.execute(
                     "INSERT INTO cards (id, nid, did, ord, mod, usn, type, queue, due, ivl, factor, reps, lapses, left, odue, odid, flags, data)
                      VALUES (?, ?, ?, ?, ?, -1, 0, 0, ?, 0, 0, 0, 0, 0, 0, 0, 0, '')",
-                    rusqlite::params![card_id, note_id, deck_id, ord, now, card_id_gen],
+                    rusqlite::params![card_id, note_id, deck_id, ord as i64, now, card_id_gen],
                 )?;
             }
         }
